chore: move python snippets to separate files
