+++ src/data/conceptAtoms.ts
@@ -1,4 +1,44 @@
 import { ConceptAtomType } from '../types';
+import codeExample1 from '../python-examples/code_example_1.py?raw';
+import codeExample2 from '../python-examples/code_example_2.py?raw';
+import codeExample3 from '../python-examples/code_example_3.py?raw';
+import codeExample4 from '../python-examples/code_example_4.py?raw';
+import codeExample5 from '../python-examples/code_example_5.py?raw';
+import codeExample6 from '../python-examples/code_example_6.py?raw';
+import codeExample7 from '../python-examples/code_example_7.py?raw';
+import codeExample8 from '../python-examples/code_example_8.py?raw';
+import codeExample9 from '../python-examples/code_example_9.py?raw';
+import codeExample10 from '../python-examples/code_example_10.py?raw';
+import codeExample11 from '../python-examples/code_example_11.py?raw';
+import codeExample12 from '../python-examples/code_example_12.py?raw';
+import codeExample13 from '../python-examples/code_example_13.py?raw';
+import codeExample14 from '../python-examples/code_example_14.py?raw';
+import codeExample15 from '../python-examples/code_example_15.py?raw';
+import codeExample16 from '../python-examples/code_example_16.py?raw';
+import codeExample17 from '../python-examples/code_example_17.py?raw';
+import codeExample18 from '../python-examples/code_example_18.py?raw';
+import codeExample19 from '../python-examples/code_example_19.py?raw';
+import codeExample20 from '../python-examples/code_example_20.py?raw';
+import codeExample21 from '../python-examples/code_example_21.py?raw';
+import codeExample22 from '../python-examples/code_example_22.py?raw';
+import codeExample23 from '../python-examples/code_example_23.py?raw';
+import codeExample24 from '../python-examples/code_example_24.py?raw';
+import codeExample25 from '../python-examples/code_example_25.py?raw';
+import codeExample26 from '../python-examples/code_example_26.py?raw';
+import codeExample27 from '../python-examples/code_example_27.py?raw';
+import codeExample28 from '../python-examples/code_example_28.py?raw';
+import solution1 from '../python-examples/solution_1.py?raw';
+import solution2 from '../python-examples/solution_2.py?raw';
+import solution3 from '../python-examples/solution_3.py?raw';
+import solution4 from '../python-examples/solution_4.py?raw';
+import solution5 from '../python-examples/solution_5.py?raw';
+import solution6 from '../python-examples/solution_6.py?raw';
+import solution7 from '../python-examples/solution_7.py?raw';
+import solution8 from '../python-examples/solution_8.py?raw';
+import solution9 from '../python-examples/solution_9.py?raw';
+import solution10 from '../python-examples/solution_10.py?raw';
+import solution11 from '../python-examples/solution_11.py?raw';
+import solution12 from '../python-examples/solution_12.py?raw';
 
 export const conceptAtoms: ConceptAtomType[] = [
   {
@@ -20,11 +60,7 @@ export const conceptAtoms: ConceptAtomType[] = [
 **Easiest Start (Online):** We'll use Replit, a free online coding environment. Just go to replit.com, create a free account, and click "Create Repl" → "Python". Name it something fun like "MyFirstBot"!
 
 **What You'll See:** You'll see a main area where you write code, and a black window on the right - that's called the "terminal" or "console." This is where your bot will "talk" back to you.`,
-            codeExample: `# This is your first line of code!
-# The # symbol means this is a "comment" - notes for humans
-print("Hello, Bot Builder!")
-
-# When you click "Run", this text will appear in the terminal`
+            codeExample: codeExample1
           },
           {
             title: 'Running Your First Line',
@@ -141,8 +177,7 @@ It's like having access to a huge library of information that your bot can 'read
           {
             title: '"Hello, World!" for Bots',
             content: `Let's start with the absolute simplest bot - one that just says something when you run it. This is like teaching your parrot its first word!`,
-            codeExample: `print("Hello! I am your first bot.")
-print("I'm excited to learn with you!")`
+            codeExample: codeExample2
           },
           {
             title: 'Getting Input (The Listener)',
@@ -151,30 +186,19 @@ print("I'm excited to learn with you!")`
 2. It pauses your program until you type something and press Enter
 
 Whatever you type gets stored in a "variable" (think of it as a labeled box) so your bot can remember it.`,
-            codeExample: `user_name = input("What's your name? ")
-print("Nice to meet you, " + user_name + "!")
-
-# The + sign here isn't for math - it's for "sticking text together"`
+            codeExample: codeExample3
           },
           {
             title: 'Basic Decision Making (The Thinker)',
             content: `Now for your bot's first "brain cell"! The \`if\` and \`else\` statements let your bot make simple choices based on what the user says.
 
 **Important:** The \`==\` asks "is this exactly equal to?" It's different from a single \`=\` which means "store this in that box."`,
-            codeExample: `user_mood = input("How are you feeling today? (Type 'good' or 'bad') ")
-
-# This is where your bot makes a decision!
-if user_mood == "good":  # Is user_mood EXACTLY "good"?
-    print("That's fantastic! Keep smiling!")
-else:  # If it's NOT "good" (meaning "bad" or anything else)
-    print("Oh no, I hope your day gets better soon!")
-
-print("Thanks for chatting with me!")`
+            codeExample: codeExample4
           },
           {
             title: 'Common Input Mistakes',
             content: `Users don't always type what you expect. If you ask for a number and they type words, Python raises a \`ValueError\`. Wrap number conversions in a \`try/except\` block so your bot can respond nicely instead of crashing.`,
-            codeExample: `try:\n    user_age = int(input("How old are you? "))\nexcept ValueError:\n    print("Please enter a number, not words!")`
+            codeExample: codeExample5
           },
           {
             title: 'Quick Tasks',
@@ -187,8 +211,7 @@ print("Thanks for chatting with me!")`
           {
             type: 'code',
             question: 'Create a bot that asks for your favorite color and then says "Wow, [color] is a beautiful color!" Try it with different colors.',
-            solution: `favorite_color = input("What's your favorite color? ")
-print("Wow, " + favorite_color + " is a beautiful color!")`,
+            solution: solution1,
             hints: [
               'Use input() to ask the question and store the answer',
               'Use print() with + to combine your message with their answer',
@@ -199,11 +222,7 @@ print("Wow, " + favorite_color + " is a beautiful color!")`,
           {
             type: 'code',
             question: 'Make a bot that asks if you like pizza and responds differently to "yes" and "no". What happens if someone types something else?',
-            solution: `pizza_answer = input("Do you like pizza? (yes/no) ")
-if pizza_answer == "yes":
-    print("Pizza is amazing! What's your favorite topping?")
-else:
-    print("That's okay, everyone has different tastes!")`,
+            solution: solution2,
             hints: [
               'Use input() to get their answer',
               'Use if/else to check if they said "yes"',
@@ -215,7 +234,7 @@ else:
           {
             type: 'code',
             question: 'Update the pizza bot to respond to "maybe" with a friendly message in addition to yes or no.',
-            solution: `pizza_answer = input("Do you like pizza? (yes/no/maybe) ")\nif pizza_answer == "yes":\n    print("Pizza is amazing! What's your favorite topping?")\nelif pizza_answer == "maybe":\n    print("It's hard to decide! Maybe try a new flavor.")\nelse:\n    print("That's okay, everyone has different tastes!")`,
+            solution: solution3,
             hints: [
               'Add an elif branch that checks for "maybe"',
               'Remember to compare the input exactly',
@@ -244,17 +263,7 @@ else:
             content: `Think of variables as labeled boxes where your bot stores information. When you write \`user_name = "Alice"\`, you're putting "Alice" in a box labeled "user_name".
 
 Your bot can look in that box anytime by using the label. This is how your bot "remembers" things!`,
-            codeExample: `# Creating variables (putting things in labeled boxes)
-user_name = "Alice"
-user_age = 25
-user_city = "New York"
-
-# Using variables (looking in the boxes)
-print("Hello " + user_name)
-print("You are " + str(user_age) + " years old")
-print("You live in " + user_city)
-
-# Note: str() converts numbers to text so we can combine them`
+            codeExample: codeExample6
           },
           {
             title: 'Different Types of Information',
@@ -266,39 +275,17 @@ print("You live in " + user_city)
 **True/False (booleans):** Special words - True, False
 
 Python is smart and figures out what type each variable is automatically!`,
-            codeExample: `# Different types of variables
-bot_name = "ChatBot"        # Text (string)
-bot_age = 1                 # Number (integer)  
-bot_height = 5.5            # Decimal (float)
-bot_is_friendly = True      # True/False (boolean)
-
-# You can check what type something is
-print(type(bot_name))       # <class 'str'>
-print(type(bot_age))        # <class 'int'>
-print(type(bot_height))     # <class 'float'>
-print(type(bot_is_friendly)) # <class 'bool'>`
+            codeExample: codeExample7
           },
           {
             title: 'Variables Can Change',
             content: `The cool thing about variables is they can change! You can put new things in the same labeled box anytime.`,
-            codeExample: `# Variables can change!
-mood = "happy"
-print("I am " + mood)
-
-mood = "excited"  # Same box, new contents
-print("Now I am " + mood)
-
-# You can even do math with number variables
-score = 10
-print("Score: " + str(score))
-
-score = score + 5  # Add 5 to the current score
-print("New score: " + str(score))`
+            codeExample: codeExample8
           },
           {
             title: 'Beware of Bad Input',
             content: `When you turn the result of \`input()\` into a number, users might type something that isn't numeric. This triggers a \`ValueError\`. Use \`try/except\` so your bot can explain the problem instead of crashing.`,
-            codeExample: `try:\n    age = int(input("How old are you? "))\nexcept ValueError:\n    print("Oops! Please enter a numeric age.")`
+            codeExample: codeExample9
           },
           {
             title: 'Quick Tasks',
@@ -311,13 +298,7 @@ print("New score: " + str(score))`
           {
             type: 'code',
             question: 'Create a bot that asks for someone\'s name and age, stores them in variables, then calculates and tells them what year they were born.',
-            solution: `name = input("What's your name? ")
-age = int(input("How old are you? "))
-
-birth_year = 2024 - age
-
-print("Hello " + name + "!")
-print("You were born in " + str(birth_year))`,
+            solution: solution4,
             hints: [
               'Use input() to get the name and age',
               'Use int() to convert the age from text to a number',
@@ -358,18 +339,7 @@ print("You were born in " + str(birth_year))`,
             content: `Your bot can make more than just yes/no decisions! With \`elif\` (short for "else if"), your bot can check multiple conditions in order.
 
 Think of it like a flowchart: "If this, do that. Otherwise, if this other thing, do that instead. Otherwise, do this default thing."`,
-            codeExample: `age = int(input("How old are you? "))
-
-if age < 13:
-    print("You're a kid! Enjoy being young!")
-elif age < 20:
-    print("You're a teenager! These are exciting years!")
-elif age < 65:
-    print("You're an adult! Hope you're enjoying life!")
-else:
-    print("You're a wise elder! Thanks for all your experience!")
-
-print("Age is just a number anyway!")`
+            codeExample: codeExample10
           },
           {
             title: 'Making Comparisons',
@@ -381,48 +351,17 @@ print("Age is just a number anyway!")`
 **>** means "greater than"
 **<=** means "less than or equal to"
 **>=** means "greater than or equal to"`,
-            codeExample: `score = int(input("What's your test score? "))
-
-if score >= 90:
-    grade = "A"
-    print("Excellent work!")
-elif score >= 80:
-    grade = "B" 
-    print("Good job!")
-elif score >= 70:
-    grade = "C"
-    print("Not bad!")
-elif score >= 60:
-    grade = "D"
-    print("You passed, but try harder next time.")
-else:
-    grade = "F"
-    print("Don't give up! You'll do better next time.")
-
-print("Your grade is: " + grade)`
+            codeExample: codeExample11
           },
           {
             title: 'Handling Different Cases',
             content: `Remember when we tried typing "YES" instead of "yes" and it didn't work? Let's fix that! The \`.lower()\` method converts text to lowercase, so "YES", "yes", and "Yes" all become "yes".`,
-            codeExample: `answer = input("Do you want to continue? (yes/no) ")
-
-# Convert to lowercase so YES, yes, Yes all work the same
-answer = answer.lower()
-
-if answer == "yes":
-    print("Great! Let's keep going!")
-elif answer == "no":
-    print("Okay, see you later!")
-else:
-    print("I didn't understand that. Please type 'yes' or 'no'.")
-
-# You can also do it in one line:
-# if answer.lower() == "yes":`
+            codeExample: codeExample12
           },
           {
             title: 'Troubleshooting User Choices',
             content: `Sometimes users type letters when you expect numbers or give unexpected answers. Wrap number conversions in \`try/except\` and provide friendly error messages when their input doesn't match what you expect.`,
-            codeExample: `try:\n    choice = int(input("Pick a number 1-3: "))\nexcept ValueError:\n    print("Numbers only! Try again.")`
+            codeExample: codeExample13
           },
           {
             title: 'Quick Tasks',
@@ -435,17 +374,7 @@ else:
           {
             type: 'code',
             question: 'Create a simple restaurant bot that asks what meal they want (breakfast/lunch/dinner) and recommends a dish for each, with a default message for anything else.',
-            solution: `meal = input("What meal are you looking for? (breakfast/lunch/dinner) ")
-meal = meal.lower()
-
-if meal == "breakfast":
-    print("I recommend pancakes with fresh berries!")
-elif meal == "lunch":
-    print("How about a delicious sandwich and soup?")
-elif meal == "dinner":
-    print("Try our famous pasta with garlic bread!")
-else:
-    print("I'm not sure about that meal. We serve breakfast, lunch, and dinner!")`,
+            solution: solution5,
             hints: [
               'Use input() to ask about the meal',
               'Use .lower() to handle different capitalizations',
@@ -457,20 +386,7 @@ else:
           {
             type: 'code',
             question: 'Make a "Magic 8-Ball" bot that asks a yes/no question, then gives a random response. Use a number input (1-3) to pick from three different answers.',
-            solution: `question = input("Ask me a yes/no question: ")
-choice = int(input("Pick a number 1-3 for your answer: "))
-
-print("\\nYou asked: " + question)
-print("The Magic 8-Ball says...")
-
-if choice == 1:
-    print("Yes, definitely!")
-elif choice == 2:
-    print("No, I don't think so.")
-elif choice == 3:
-    print("Maybe... ask again later.")
-else:
-    print("Please pick 1, 2, or 3 next time!")`,
+            solution: solution6,
             hints: [
               'Ask for their question first (just for fun)',
               'Get a number from 1-3 using int(input())',
@@ -512,17 +428,7 @@ else:
             content: `Imagine you want your bot to count to 10. You could write 10 print statements... or you could use a loop! Loops let your bot repeat actions without you having to write the same code over and over.
 
 There are two main types: **for loops** (when you know how many times) and **while loops** (when you want to keep going until something happens).`,
-            codeExample: `# The hard way (don't do this!)
-print("1")
-print("2")
-print("3")
-# ... this would take forever!
-
-# The smart way with a for loop
-for number in range(1, 11):  # 1 to 10
-    print(number)
-
-print("Done counting!")`
+            codeExample: codeExample14
           },
           {
             title: 'For Loops: When You Know How Many Times',
@@ -531,70 +437,24 @@ print("Done counting!")`
 **range(5)** gives you: 0, 1, 2, 3, 4 (5 numbers starting from 0)
 **range(1, 6)** gives you: 1, 2, 3, 4, 5 (from 1 to 5)
 **range(0, 10, 2)** gives you: 0, 2, 4, 6, 8 (every 2nd number)`,
-            codeExample: `# Greet the user multiple times
-for i in range(3):
-    print("Hello there! This is greeting number " + str(i + 1))
-
-print()  # Empty line for spacing
-
-# Countdown
-print("Rocket launch countdown:")
-for count in range(10, 0, -1):  # Count backwards from 10 to 1
-    print(str(count) + "...")
-print("Blast off! 🚀")`
+            codeExample: codeExample15
           },
           {
             title: 'While Loops: Keep Going Until...',
             content: `While loops keep repeating as long as a condition is True. They're perfect for making your bot keep asking questions until the user wants to quit.
 
 **Important:** Make sure your while loop can eventually stop, or it will run forever! Always include a way for the condition to become False.`,
-            codeExample: `# Keep asking until they say "quit"
-print("I'm a chatty bot! Type 'quit' when you want to stop.")
-
-while True:  # This means "keep going forever"
-    user_input = input("Say something to me: ")
-    
-    if user_input.lower() == "quit":
-        print("Goodbye! It was fun chatting!")
-        break  # This stops the loop
-    else:
-        print("You said: " + user_input)
-        print("That's interesting! Tell me more.")
-
-print("Thanks for chatting with me!")`
+            codeExample: codeExample16
           },
           {
             title: 'Combining Loops with Choices',
             content: `The real magic happens when you combine loops with if/else statements. Now your bot can have ongoing conversations and make different responses based on what the user says!`,
-            codeExample: `print("Welcome to the Mood Bot! I'll help track your mood.")
-print("Type 'done' when you want to finish.")
-
-mood_count = 0
-
-while True:
-    mood = input("\\nHow are you feeling right now? ")
-    
-    if mood.lower() == "done":
-        print("\\nThanks for sharing! You told me about " + str(mood_count) + " moods.")
-        break
-    
-    mood_count = mood_count + 1  # Count how many moods they shared
-    
-    if mood.lower() == "happy":
-        print("That's wonderful! Keep that positive energy!")
-    elif mood.lower() == "sad":
-        print("I'm sorry you're feeling down. Tomorrow is a new day!")
-    elif mood.lower() == "excited":
-        print("Your excitement is contagious! What's got you so pumped?")
-    else:
-        print("Thanks for sharing that you're feeling " + mood + ".")
-    
-    print("Feelings are important to acknowledge.")`
+            codeExample: codeExample17
           },
           {
             title: 'Avoiding Infinite Loops',
             content: `If your loop expects a number and someone types text, Python throws a \`ValueError\` and the loop stops. Use \`try/except\` inside the loop to catch mistakes and continue asking.`,
-            codeExample: `while True:\n    try:\n        guess = int(input("Your guess: "))\n    except ValueError:\n        print("That's not a number!")\n        continue\n    # continue with your checks...`
+            codeExample: codeExample18
           },
           {
             title: 'Quick Tasks',
@@ -607,20 +467,7 @@ while True:
           {
             type: 'code',
             question: 'Create a "Compliment Bot" that asks for the user\'s name, then gives them 3 different compliments using a for loop.',
-            solution: `name = input("What's your name? ")
-
-compliments = [
-    "You have a great sense of humor!",
-    "You're really good at learning new things!",
-    "You have a wonderful personality!"
-]
-
-print("\\nHere are some compliments for you, " + name + ":")
-
-for i in range(3):
-    print(str(i + 1) + ". " + compliments[i])
-
-print("\\nYou're awesome, " + name + "! 😊")`,
+            solution: solution7,
             hints: [
               'Ask for their name first',
               'Create a list of compliments (we\'ll learn more about lists later)',
@@ -631,23 +478,7 @@ print("\\nYou're awesome, " + name + "! 😊")`,
           {
             type: 'code',
             question: 'Make a "Guessing Game" bot that thinks of a number (let\'s say 7) and keeps asking the user to guess until they get it right.',
-            solution: `secret_number = 7
-print("I'm thinking of a number between 1 and 10!")
-print("Can you guess what it is?")
-
-while True:
-    guess = int(input("Your guess: "))
-    
-    if guess == secret_number:
-        print("🎉 Congratulations! You got it!")
-        print("The number was " + str(secret_number))
-        break
-    elif guess < secret_number:
-        print("Too low! Try a higher number.")
-    else:
-        print("Too high! Try a lower number.")
-
-print("Thanks for playing!")`,
+            solution: solution8,
             hints: [
               'Set a secret number at the start',
               'Use while True: to keep asking for guesses',
@@ -689,20 +520,7 @@ print("Thanks for playing!")`,
             content: `Remember our library analogy? An API is like a librarian who can instantly get you any book (information) you ask for. Instead of your bot making up jokes or weather data, it can ask specialized services for that information.
 
 **The best part:** Many APIs are completely free to use for learning! We'll start with a super simple one that gives us dad jokes.`,
-            codeExample: `# This is what we're building towards:
-# A bot that can get fresh jokes from the internet!
-
-import requests  # This is our "library card" for talking to websites
-
-def get_dad_joke():
-    # Ask the joke service for a joke
-    response = requests.get("https://icanhazdadjoke.com/", 
-                          headers={"Accept": "application/json"})
-    joke_data = response.json()
-    return joke_data["joke"]
-
-# Now our bot can tell internet jokes!
-print(get_dad_joke())`
+            codeExample: codeExample19
           },
           {
             title: 'Installing the Requests Library',
@@ -713,12 +531,7 @@ print(get_dad_joke())`
 \`pip install requests\`
 
 **What this does:** \`pip\` is Python's tool for adding new features. \`install requests\` gets the "requests" library and adds it to your Python setup.`,
-            codeExample: `# First, we need to import (bring in) the requests library
-import requests
-
-# Test if it's working
-print("Requests library is ready!")
-print("Now we can talk to websites and APIs!")`
+            codeExample: codeExample20
           },
           {
             title: 'Your First API Call',
@@ -730,58 +543,17 @@ print("Now we can talk to websites and APIs!")`
 4. **joke_data["joke"]** - Your bot picks out just the joke part
 
 It's like ordering at a restaurant: you ask for a specific dish, they bring you a whole tray, and you pick out what you want to eat.`,
-            codeExample: `import requests
-
-# Step 1: Ask the API for a joke
-print("Asking the internet for a joke...")
-response = requests.get("https://icanhazdadjoke.com/", 
-                       headers={"Accept": "application/json"})
-
-# Step 2: Convert the response to Python data
-joke_data = response.json()
-
-# Step 3: Extract just the joke
-joke = joke_data["joke"]
-
-# Step 4: Tell the joke!
-print("Here's your joke:")
-print(joke)`
+            codeExample: codeExample21
           },
           {
             title: 'Building a Joke Bot',
             content: `Now let's put it all together into a real bot that can tell jokes on demand! We'll combine everything we've learned: loops, user input, and APIs.`,
-            codeExample: `import requests
-
-def get_dad_joke():
-    """This function gets a joke from the internet"""
-    try:
-        response = requests.get("https://icanhazdadjoke.com/", 
-                               headers={"Accept": "application/json"})
-        joke_data = response.json()
-        return joke_data["joke"]
-    except:
-        return "Sorry, I couldn't get a joke right now. Maybe the internet is having a bad day!"
-
-# Main bot program
-print("🤖 Welcome to JokeBot! I get my jokes fresh from the internet!")
-print("Type 'joke' for a joke, or 'quit' to exit.")
-
-while True:
-    user_input = input("\\nWhat would you like? ").lower()
-    
-    if user_input == "quit":
-        print("Thanks for laughing with me! Goodbye! 😄")
-        break
-    elif user_input == "joke":
-        print("\\n" + get_dad_joke())
-        print("\\n😂 Hope that made you smile!")
-    else:
-        print("I only know how to tell jokes! Type 'joke' or 'quit'.")`
+            codeExample: codeExample22
           },
           {
             title: 'Handling API Errors',
             content: `APIs sometimes fail or your internet connection might drop. When \`requests.get\` can't reach the server, it raises an exception. Check the status code and wrap the call in \`try/except\` so your bot can display a friendly message instead of crashing.`,
-            codeExample: `def safe_joke():\n    try:\n        r = requests.get("https://icanhazdadjoke.com/", headers={"Accept": "application/json"})\n        if r.status_code == 200:\n            return r.json()["joke"]\n        return "API returned an error."\n    except requests.exceptions.RequestException:\n        return "Network error. Try again later."`
+            codeExample: codeExample23
           },
           {
             title: 'Quick Tasks',
@@ -794,32 +566,7 @@ while True:
           {
             type: 'code',
             question: 'Modify the joke bot to keep track of how many jokes it has told, and display the count each time.',
-            solution: `import requests
-
-def get_dad_joke():
-    try:
-        response = requests.get("https://icanhazdadjoke.com/", 
-                               headers={"Accept": "application/json"})
-        joke_data = response.json()
-        return joke_data["joke"]
-    except:
-        return "Sorry, I couldn't get a joke right now!"
-
-print("🤖 Welcome to JokeBot with Counter!")
-joke_count = 0
-
-while True:
-    user_input = input("\\nType 'joke' for a joke, or 'quit' to exit: ").lower()
-    
-    if user_input == "quit":
-        print(f"Thanks for laughing! I told you {joke_count} jokes today! 😄")
-        break
-    elif user_input == "joke":
-        joke_count += 1
-        print(f"\\nJoke #{joke_count}:")
-        print(get_dad_joke())
-    else:
-        print("Type 'joke' or 'quit'!")`,
+            solution: solution9,
             hints: [
               'Create a joke_count variable before the loop',
               'Add 1 to joke_count each time you tell a joke',
@@ -831,28 +578,7 @@ while True:
           {
             type: 'code',
             question: 'Create a "Fact Bot" using the Numbers API (http://numbersapi.com/random). Make it get random number facts instead of jokes.',
-            solution: `import requests
-
-def get_number_fact():
-    try:
-        response = requests.get("http://numbersapi.com/random")
-        return response.text  # This API returns plain text, not JSON
-    except:
-        return "Sorry, I couldn't get a fact right now!"
-
-print("🧠 Welcome to FactBot! I know amazing number facts!")
-
-while True:
-    user_input = input("\\nType 'fact' for a number fact, or 'quit' to exit: ").lower()
-    
-    if user_input == "quit":
-        print("Thanks for learning with me! Stay curious! 🤓")
-        break
-    elif user_input == "fact":
-        print("\\nHere's a cool number fact:")
-        print(get_number_fact())
-    else:
-        print("Type 'fact' or 'quit'!")`,
+            solution: solution10,
             hints: [
               'Use http://numbersapi.com/random as the URL',
               'This API returns plain text, so use response.text instead of response.json()',
@@ -903,150 +629,22 @@ while True:
 - What kind of responses it gives  
 - What APIs it uses
 - How it handles errors`,
-            codeExample: `# Example: A friendly, encouraging bot personality
-def bot_greeting():
-    return "Hey there, friend! 😊 I'm here to help and chat!"
-
-def bot_encouragement():
-    encouragements = [
-        "You're doing great!",
-        "Keep up the awesome work!",
-        "I believe in you!",
-        "You've got this!"
-    ]
-    import random
-    return random.choice(encouragements)`
+            codeExample: codeExample24
           },
           {
             title: 'Adding Multiple Features with elif',
             content: `Instead of just telling jokes, let's make a bot that can do several things! We'll use \`elif\` to check for different commands.`,
-            codeExample: `import requests
-import random
-
-def get_dad_joke():
-    try:
-        response = requests.get("https://icanhazdadjoke.com/", 
-                               headers={"Accept": "application/json"})
-        return response.json()["joke"]
-    except:
-        return "Sorry, joke service is down!"
-
-def get_encouragement():
-    encouragements = [
-        "You're amazing just as you are!",
-        "Every day is a new opportunity!",
-        "You have the power to make today great!",
-        "Believe in yourself - I do!"
-    ]
-    return random.choice(encouragements)
-
-def get_fun_fact():
-    try:
-        response = requests.get("http://numbersapi.com/random")
-        return response.text
-    except:
-        return "Here's a fact: You're awesome for learning to code!"
-
-# Multi-feature bot
-print("🤖 Hi! I'm your Multi-Bot! I can do lots of things!")
-print("Commands: joke, encourage, fact, help, quit")
-
-while True:
-    command = input("\\nWhat can I do for you? ").lower()
-    
-    if command == "quit":
-        print("Goodbye! You're going to do amazing things! 🌟")
-        break
-    elif command == "joke":
-        print("😄 " + get_dad_joke())
-    elif command == "encourage":
-        print("💪 " + get_encouragement())
-    elif command == "fact":
-        print("🧠 " + get_fun_fact())
-    elif command == "help":
-        print("I can: tell jokes, give encouragement, share facts!")
-        print("Just type: joke, encourage, fact, help, or quit")
-    else:
-        print("🤔 I don't know that command. Type 'help' to see what I can do!")`
+            codeExample: codeExample25
           },
           {
             title: 'Adding Randomness and Variety',
             content: `To make your bot feel more natural, add some randomness! Instead of always saying the same thing, your bot can pick from different responses.`,
-            codeExample: `import random
-
-# Lists of different responses
-greetings = [
-    "Hello there!",
-    "Hey, what's up?",
-    "Greetings, human!",
-    "Hi there, friend!",
-    "Welcome back!"
-]
-
-goodbyes = [
-    "See you later!",
-    "Goodbye for now!",
-    "Until next time!",
-    "Take care!",
-    "Catch you later!"
-]
-
-confused_responses = [
-    "I'm not sure what you mean.",
-    "Could you try that again?",
-    "I didn't quite understand that.",
-    "Hmm, I'm confused. Can you rephrase?",
-    "That's not something I know how to do."
-]
-
-# Using random responses
-print(random.choice(greetings))
-
-# Later in your bot...
-if command not in ["joke", "encourage", "fact", "help", "quit"]:
-    print(random.choice(confused_responses))`
+            codeExample: codeExample26
           },
           {
             title: 'Remembering User Information',
             content: `Make your bot more personal by remembering things about the user! We can store information in variables and use it throughout the conversation.`,
-            codeExample: `import random
-
-# Store user information
-user_name = ""
-favorite_color = ""
-mood_today = ""
-
-print("🤖 Hello! I'm PersonalBot. I like getting to know people!")
-
-# Get to know the user
-user_name = input("What's your name? ")
-print(f"Nice to meet you, {user_name}!")
-
-favorite_color = input("What's your favorite color? ")
-print(f"{favorite_color} is a beautiful color, {user_name}!")
-
-mood_today = input("How are you feeling today? ")
-print(f"Thanks for sharing that you're feeling {mood_today}.")
-
-print(f"\\nOkay {user_name}, now I know a bit about you!")
-print("Type 'chat' to talk, or 'quit' to leave.")
-
-while True:
-    command = input("\\n> ").lower()
-    
-    if command == "quit":
-        print(f"Goodbye {user_name}! I hope your {favorite_color} day gets even better!")
-        break
-    elif command == "chat":
-        responses = [
-            f"So {user_name}, tell me more about yourself!",
-            f"I remember you like {favorite_color}. What else do you enjoy?",
-            f"Since you're feeling {mood_today}, what usually makes you happy?",
-            f"You seem really nice, {user_name}. What's your favorite hobby?"
-        ]
-        print(random.choice(responses))
-    else:
-        print(f"Just type 'chat' to talk with me, {user_name}, or 'quit' to leave!")`
+            codeExample: codeExample27
           },
           {
             title: 'Quick Tasks',
@@ -1059,60 +657,7 @@ while True:
           {
             type: 'code',
             question: 'Create a "Study Buddy Bot" that can give study tips, motivational quotes, and quiz the user with simple math problems. Include at least 3 different features.',
-            solution: `import random
-
-def get_study_tip():
-    tips = [
-        "Take breaks every 25 minutes - your brain needs rest!",
-        "Study in a quiet place without distractions.",
-        "Teach someone else what you learned - it helps you remember!",
-        "Use flashcards for memorizing facts.",
-        "Get enough sleep - tired brains don't learn well!"
-    ]
-    return random.choice(tips)
-
-def get_motivation():
-    quotes = [
-        "The expert in anything was once a beginner!",
-        "Every mistake is a learning opportunity!",
-        "You're building your future with every study session!",
-        "Knowledge is power - keep gaining that power!",
-        "Small progress is still progress!"
-    ]
-    return random.choice(quotes)
-
-def math_quiz():
-    num1 = random.randint(1, 10)
-    num2 = random.randint(1, 10)
-    answer = num1 + num2
-    
-    user_answer = int(input(f"Quick math: {num1} + {num2} = ? "))
-    
-    if user_answer == answer:
-        return "🎉 Correct! Your brain is working great!"
-    else:
-        return f"Not quite! The answer is {answer}. Keep practicing!"
-
-print("📚 Hi! I'm StudyBot, your learning companion!")
-print("Commands: tip, motivate, quiz, help, quit")
-
-while True:
-    command = input("\\nHow can I help your studies? ").lower()
-    
-    if command == "quit":
-        print("Happy studying! Remember, every expert was once a beginner! 🌟")
-        break
-    elif command == "tip":
-        print("💡 " + get_study_tip())
-    elif command == "motivate":
-        print("💪 " + get_motivation())
-    elif command == "quiz":
-        print("🧮 " + math_quiz())
-    elif command == "help":
-        print("I can give study tips, motivation, and quiz you!")
-        print("Commands: tip, motivate, quiz, help, quit")
-    else:
-        print("Type 'help' to see what I can do for your studies!")`,
+            solution: solution11,
             hints: [
               'Create separate functions for each feature',
               'Use lists and random.choice() for variety',
@@ -1124,74 +669,7 @@ while True:
           {
             type: 'code',
             question: 'Design a "Mood Tracker Bot" that asks about the user\'s mood, remembers previous moods in a list, and can show a summary of how they\'ve been feeling.',
-            solution: `import random
-
-# Store mood history
-mood_history = []
-user_name = ""
-
-def add_mood(mood):
-    mood_history.append(mood)
-    responses = [
-        f"Thanks for sharing that you're feeling {mood}.",
-        f"I've noted that you're {mood} today.",
-        f"Feeling {mood} is totally valid.",
-        f"I appreciate you telling me you're {mood}."
-    ]
-    return random.choice(responses)
-
-def show_mood_summary():
-    if not mood_history:
-        return "You haven't shared any moods with me yet!"
-    
-    summary = f"You've shared {len(mood_history)} moods with me:\\n"
-    for i, mood in enumerate(mood_history, 1):
-        summary += f"{i}. {mood}\\n"
-    
-    # Find most common mood
-    if mood_history:
-        most_common = max(set(mood_history), key=mood_history.count)
-        summary += f"\\nYou seem to feel '{most_common}' quite often!"
-    
-    return summary
-
-def get_mood_advice(mood):
-    if mood.lower() in ["sad", "down", "depressed", "upset"]:
-        return "Remember: tough times don't last, but tough people do! 💪"
-    elif mood.lower() in ["happy", "great", "amazing", "excited"]:
-        return "That's wonderful! Keep spreading those positive vibes! ✨"
-    elif mood.lower() in ["tired", "exhausted", "sleepy"]:
-        return "Rest is important! Make sure you're getting enough sleep. 😴"
-    else:
-        return "All feelings are valid. Thanks for sharing with me! 🤗"
-
-print("🌈 Hello! I'm MoodBot. I help you track and understand your feelings.")
-user_name = input("What's your name? ")
-print(f"Nice to meet you, {user_name}! Let's start tracking your moods.")
-print("Commands: mood, summary, advice, help, quit")
-
-while True:
-    command = input(f"\\nHi {user_name}, what would you like to do? ").lower()
-    
-    if command == "quit":
-        print(f"Goodbye {user_name}! Remember to check in with your feelings regularly! 💝")
-        break
-    elif command == "mood":
-        current_mood = input("How are you feeling right now? ")
-        print(add_mood(current_mood))
-    elif command == "summary":
-        print("📊 " + show_mood_summary())
-    elif command == "advice":
-        if mood_history:
-            latest_mood = mood_history[-1]
-            print("💭 " + get_mood_advice(latest_mood))
-        else:
-            print("Share a mood with me first, then I can give advice!")
-    elif command == "help":
-        print("I can track your moods and give summaries!")
-        print("Commands: mood (add a mood), summary (see all moods), advice (get mood advice), help, quit")
-    else:
-        print("Type 'help' to see what I can do!")`,
+            solution: solution12,
             hints: [
               'Use a list to store all the moods they share',
               'Create functions for adding moods, showing summaries, and giving advice',
@@ -1283,40 +761,7 @@ This is a creative exercise. Enjoy the process of bringing your bot to life. Eve
           {
             title: 'Quick Reference: Your Bot Building Toolkit',
             content: `Keep this handy while building your challenge bot:`,
-            codeExample: `# Essential Bot Structure
-import requests
-import random
-
-# Bot personality
-def bot_greeting():
-    return "Hello! I'm [YourBotName]!"
-
-# API function
-def get_api_data():
-    try:
-        response = requests.get("YOUR_API_URL")
-        return response.json()  # or response.text
-    except:
-        return "Sorry, couldn't get that info right now!"
-
-# Main conversation loop
-print(bot_greeting())
-print("Commands: [list your commands], help, quit")
-
-while True:
-    user_input = input("\\n> ").lower()
-    
-    if user_input == "quit":
-        print("Goodbye! [Encouraging message]")
-        break
-    elif user_input == "help":
-        print("I can: [list features]")
-    elif user_input == "[your_command]":
-        print(get_api_data())
-    else:
-        print("Type 'help' to see what I can do!")
-
-# Remember: Test each piece before adding the next!`
+            codeExample: codeExample28
           },
           {
             title: 'Where to Go Next (For the Curious!)',
