+++ src/python-examples/code_example_1.py
@@ -1,0 +1,5 @@
+# This is your first line of code!
+# The # symbol means this is a "comment" - notes for humans
+print("Hello, Bot Builder!")
+
+# When you click "Run", this text will appear in the terminal
