+++ src/python-examples/code_example_10.py
@@ -1,0 +1,12 @@
+age = int(input("How old are you? "))
+
+if age < 13:
+    print("You're a kid! Enjoy being young!")
+elif age < 20:
+    print("You're a teenager! These are exciting years!")
+elif age < 65:
+    print("You're an adult! Hope you're enjoying life!")
+else:
+    print("You're a wise elder! Thanks for all your experience!")
+
+print("Age is just a number anyway!")
