+++ src/python-examples/code_example_13.py
@@ -1,0 +1,1 @@
+try:\n    choice = int(input("Pick a number 1-3: "))\nexcept ValueError:\n    print("Numbers only! Try again.")
