+++ src/python-examples/code_example_14.py
@@ -1,0 +1,11 @@
+# The hard way (don't do this!)
+print("1")
+print("2")
+print("3")
+# ... this would take forever!
+
+# The smart way with a for loop
+for number in range(1, 11):  # 1 to 10
+    print(number)
+
+print("Done counting!")
