+++ src/python-examples/code_example_18.py
@@ -1,0 +1,1 @@
+while True:\n    try:\n        guess = int(input("Your guess: "))\n    except ValueError:\n        print("That's not a number!")\n        continue\n    # continue with your checks...
