+++ src/python-examples/code_example_2.py
@@ -1,0 +1,2 @@
+print("Hello! I am your first bot.")
+print("I'm excited to learn with you!")
