+++ src/python-examples/code_example_20.py
@@ -1,0 +1,6 @@
+# First, we need to import (bring in) the requests library
+import requests
+
+# Test if it's working
+print("Requests library is ready!")
+print("Now we can talk to websites and APIs!")
