+++ src/python-examples/code_example_28.py
@@ -1,0 +1,34 @@
+# Essential Bot Structure
+import requests
+import random
+
+# Bot personality
+def bot_greeting():
+    return "Hello! I'm [YourBotName]!"
+
+# API function
+def get_api_data():
+    try:
+        response = requests.get("YOUR_API_URL")
+        return response.json()  # or response.text
+    except:
+        return "Sorry, couldn't get that info right now!"
+
+# Main conversation loop
+print(bot_greeting())
+print("Commands: [list your commands], help, quit")
+
+while True:
+    user_input = input("\\n> ").lower()
+    
+    if user_input == "quit":
+        print("Goodbye! [Encouraging message]")
+        break
+    elif user_input == "help":
+        print("I can: [list features]")
+    elif user_input == "[your_command]":
+        print(get_api_data())
+    else:
+        print("Type 'help' to see what I can do!")
+
+# Remember: Test each piece before adding the next!
