+++ src/python-examples/code_example_3.py
@@ -1,0 +1,4 @@
+user_name = input("What's your name? ")
+print("Nice to meet you, " + user_name + "!")
+
+# The + sign here isn't for math - it's for "sticking text together"
