+++ src/python-examples/code_example_5.py
@@ -1,0 +1,1 @@
+try:\n    user_age = int(input("How old are you? "))\nexcept ValueError:\n    print("Please enter a number, not words!")
