+++ src/python-examples/code_example_6.py
@@ -1,0 +1,11 @@
+# Creating variables (putting things in labeled boxes)
+user_name = "Alice"
+user_age = 25
+user_city = "New York"
+
+# Using variables (looking in the boxes)
+print("Hello " + user_name)
+print("You are " + str(user_age) + " years old")
+print("You live in " + user_city)
+
+# Note: str() converts numbers to text so we can combine them
