+++ src/python-examples/code_example_7.py
@@ -1,0 +1,11 @@
+# Different types of variables
+bot_name = "ChatBot"        # Text (string)
+bot_age = 1                 # Number (integer)  
+bot_height = 5.5            # Decimal (float)
+bot_is_friendly = True      # True/False (boolean)
+
+# You can check what type something is
+print(type(bot_name))       # <class 'str'>
+print(type(bot_age))        # <class 'int'>
+print(type(bot_height))     # <class 'float'>
+print(type(bot_is_friendly)) # <class 'bool'>
