+++ src/python-examples/code_example_8.py
@@ -1,0 +1,13 @@
+# Variables can change!
+mood = "happy"
+print("I am " + mood)
+
+mood = "excited"  # Same box, new contents
+print("Now I am " + mood)
+
+# You can even do math with number variables
+score = 10
+print("Score: " + str(score))
+
+score = score + 5  # Add 5 to the current score
+print("New score: " + str(score))
