+++ src/python-examples/code_example_9.py
@@ -1,0 +1,1 @@
+try:\n    age = int(input("How old are you? "))\nexcept ValueError:\n    print("Oops! Please enter a numeric age.")
