+++ src/python-examples/solution_1.py
@@ -1,0 +1,2 @@
+favorite_color = input("What's your favorite color? ")
+print("Wow, " + favorite_color + " is a beautiful color!")
