+++ src/python-examples/solution_2.py
@@ -1,0 +1,5 @@
+pizza_answer = input("Do you like pizza? (yes/no) ")
+if pizza_answer == "yes":
+    print("Pizza is amazing! What's your favorite topping?")
+else:
+    print("That's okay, everyone has different tastes!")
