+++ src/python-examples/solution_4.py
@@ -1,0 +1,7 @@
+name = input("What's your name? ")
+age = int(input("How old are you? "))
+
+birth_year = 2024 - age
+
+print("Hello " + name + "!")
+print("You were born in " + str(birth_year))
